dashboard/requirements.txt
streamlit
pandas
numpy
paho-mqtt
requests
